change how we deal w/ incompatible node versions
Previously we reported an error on registration and never scheduled any polling
for breakpoints. The Debug API would like us to report the error at set-break-
point time instead. The problem is that the debuggee (since it is in error
state) times out on the server and is no longer active. This means that we
cannot reliably get the error reported on the pantheon UI.

The proposed solution is to report the error at set-breakpoint time instead.