Heuristic for resolving ambiguous file paths
This heuristic will correctly find scripts if their name is unique in
the executing project regardless of path to handle the case where the
app was deployed to a different path than that used in development.

Additionally, if the environment variable GCLOUD_DEBUG_GIT_ROOT is
provided, breakpoint paths will be modified to assume the project root
is the value provided in the variable.

Progress on #51