Test for source context
This ensures that the source context read from the source-contexts.json
file in the working directory is passed to the api when the debuggee
registers.

Also look for a file named source-contexts.json instead of
source-context.json since the gcloud tool now generates this file name
by default and deprecated the ability to use other names.

Fixes #15.