[SQUASH] Init commit to prioritize capturing exps
In particular, now if a captured value has a `length` attribute,
the value of that attribute is shown in the status field.

Also, the length of a watched expression is now bounded by
`config.maxDataSize` instead of `config.maxProperties`.

Unit tests still need to be addressed, and more fine grained
control of the total length of the data captured needs to be
added.