[SQUASH] Init commit
Now the cloud debugger is started by explicitly calling the
newly exposed `start()` method that accepts an optional config
object.  Currently if the `start()` method is not called
within 5 seconds it is automatically called and the user is
issued a message detailing that the `start()` method will need
to be explicitly called in future releases.