[SQUASH] Fix a bug testing the log level
If the GCLOUD_DEBUG_LOGLEVEL environment variable is set then
its value will be used as the log level instead of the value
specified in the supplied configuration file.  However, the
unit test had the value of GCLOUD_DEBUG_LOGLEVEL and the value
in the config file both set to the same value.  Thus, the test
wouldn't verify if the environment variable was actually taking
precedence.