Add improved support for transpiled code (#159)
This PR includes the new SourceMapper that is responsible for processing sourcemap files and understanding which files in a project are the input files and output files to a transpilation process.  

This adds robustness to handling breakpoints set in source files used as the input to Babel, Typescript, and Coffeescript transpilation.

This pull request also addresses the bug [#153](https://github.com/GoogleCloudPlatform/cloud-debug-nodejs/issues/153) "Off-by-one-line error in transpiled breakpoint".