Prioritize capturing expressions (#162)
Now the collection of the contents of watched expressions is given priority over other variables such as locals.  Further, the contents of watched expressions are shown completely provided that this can be done within the limits of `config.capture.maxDataSize`.  Otherwise, a message is displayed describing why the entire contents of the watched expression cannot be displayed.

For example, if a breakpoint is hit with a watched expression that is an array (even if that array contains more than `config.capture.maxProperties` elements), the entire contents of the array will be collected provided that the limit `config.capture.maxDataSize` is not reached.  Then, other variables such as locals will be collected (provided that there is enough space to collect that information as specified by `config.capture.maxDataSize`).