API changes bring us closer to `google-cloud` (#180)
This change adapts the primary interface of this module to align it
with the API exposed by all the other Google Cloud API Client
Libraries available via the `google-cloud` modules [1].

The primary change is that we return a Debug construtor as the return
value from the module. This, in the future, will provide a Client
Library interface to the Debug APIs. This can be used for programatic
access to list or modify breakpoints, logpoints etc. The constructor,
like other API clients, accepts authentication and other general
configuration.

An instantiated debug object exposes a function `startAgent` that can
be used to activate the process-wide agent that enables the debugging
of an application using the Stackdriver Debug UI in the Google Cloud
console.

[1]: https://github.com/GoogleCloudPlatform/google-cloud-node