Remove duplicate isDisabled logic from controller
The debuglet also has the same isDisabled check. Arguably the
controller shouldn't be doing this check at all because we do get a
valid debuggee.id back from the service and we should faithfully
report that to the client of the Controller interface. Business logic
to deal with a disabled debuggee belongs in the agent.