Replace v8debugapi to inspector API.
This commit implements V8 inspector API.
There are some remaining issues not solves.
1. isNative is not implemented as inspector API does not provide, and
the work around method is not sufficient.
2. in state.ts. function evaluate has been removed as functions calling
evaluate in state.ts and v8debugapi.ts has different usage with the
inspector debugger method 'eveluateOnCallFrame'. Potential refactoring
can be done in the future.
3. As most methods in inspector API are asynchronous, many methods in
state.ts have to be changed to asynchronous.
4. In the locals of callFrame in apiType.breakpoint, a 'context' field
needs to be added in the future to reflect this pointer.