Remove args in state.ts#resolveLocalList (#331)
Though cloud debugger has args in the proto, args is used to identify
passing variables to the functions. Locals, on the other hand is the
local variables defined within the function. However, stackdriver
debugger UI displays them in the same place with the order of passing
variables prior to local variables. In our case, debugger's call frame
will have the same effects of ordering passing variables prior to local
variables. We can sufficiently merge args and locals in one place.
That's the reason of removing args in the code.