fix: Remove the `process` warning listener (#376)
Node 8.5.0 introduced a change where a warning message discussing
the use of async traces would be printed when using the debug
agent with the inspector protocol.  At that time, a listener was
added that listened to that warning and also printed a message
stating that the warning does not impact the debug agent.

At this point, the underlying issue causing the warning message to
appear has been fixed in Node, and, as such, the associated warning
listener in the debug agent has be removed.