chore: generate key.json.enc using node:8 image
I think the previous key cannot be properly decrypted in CircleCI
because it was generated using a version of openssl different from
the version used during decryption.

The key in this PR was generated from within a node:8 image which
is the same image used by the system and samples tests, to make
sure the same version of openssl is used.