chore: merge `options.debug` and `options` (#718)
If the `options` object given to the `start()` function has a
debug property then, previously, the options used would be
`options.debug`.  Thus anything set in `options` itself, such
as the `projectId` would not be used by the agent.  This
change suggests using the merger of `options` and `options.debug`.